Tugas 27

+++ routes/web.php
@@ -24,52 +24,53 @@ use App\Http\Controllers\DashboardController;
 // Route::middleware(['auth',])->group(function () {
 //Kategori
 Route::get('kategori', [KategoriController::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('kategori/create', [KategoriController::class, 'create'])->middleware(['admin']);
-Route::post('kategori/store', [KategoriController::class, 'store'])->name('kategori.store')->middleware(['admin']);
-Route::get('kategori/{id}/edit', [KategoriController::class, 'edit'])->middleware(['admin']);
-Route::put('kategori/{id}', [KategoriController::class, 'update'])->middleware(['admin']);
-Route::delete('kategori/{id}', [KategoriController::class, 'destroy'])->middleware(['admin']);
+Route::get('kategori/create', [KategoriController::class, 'create'])->middleware(['auth','admin']);
+Route::post('kategori/store', [KategoriController::class, 'store'])->name('kategori.store')->middleware(['auth','admin']);
+Route::get('kategori/{id}/edit', [KategoriController::class, 'edit'])->middleware(['auth','admin']);
+Route::put('kategori/{id}', [KategoriController::class, 'update'])->middleware(['auth','admin']);
+Route::delete('kategori/{id}', [KategoriController::class, 'destroy'])->middleware(['auth','admin']);
 
 //Produk
 Route::get('produks', [Produk1Controller::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('produks/create', [Produk1Controller::class, 'create'])->middleware(['admin']);
-Route::post('produks/store', [Produk1Controller::class, 'store'])->name('produk.store')->middleware(['admin']);
-Route::get('produks/{id}/edit', [Produk1Controller::class, 'edit'])->middleware(['admin']);
-Route::put('produks/{id}', [Produk1Controller::class, 'update'])->middleware(['admin']);
-Route::delete('produks/{id}', [Produk1Controller::class, 'destroy'])->middleware(['admin']);
+Route::get('produks/create', [Produk1Controller::class, 'create'])->middleware(['auth','admin']);
+Route::post('produks/store', [Produk1Controller::class, 'store'])->name('produk.store')->middleware(['auth','admin']);
+Route::get('produks/{id}/edit', [Produk1Controller::class, 'edit'])->middleware(['auth','admin']);
+Route::put('produks/{id}', [Produk1Controller::class, 'update'])->middleware(['auth','admin']);
+Route::delete('produks/{id}', [Produk1Controller::class, 'destroy'])->middleware(['auth','admin']);
+Route::get('produks/{id}', [Produk1Controller::class, 'show'])->name('produk.show')->middleware(['auth']);
 
 //Pengguna
 Route::get('pengguna', [PenggunaController::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('pengguna/create', [PenggunaController::class, 'create'])->middleware(['admin']);
-Route::post('pengguna/store', [PenggunaController::class, 'store'])->name('pengguna.store')->middleware(['admin']);
-Route::get('pengguna/{id}/edit', [PenggunaController::class, 'edit'])->middleware(['admin']);
-Route::put('pengguna/{id}', [PenggunaController::class, 'update'])->middleware(['admin']);
-Route::delete('pengguna/{id}', [PenggunaController::class, 'destroy'])->middleware(['admin']);
+Route::get('pengguna/create', [PenggunaController::class, 'create'])->middleware(['auth','admin']);
+Route::post('pengguna/store', [PenggunaController::class, 'store'])->name('pengguna.store')->middleware(['auth','admin']);
+Route::get('pengguna/{id}/edit', [PenggunaController::class, 'edit'])->middleware(['auth','admin']);
+Route::put('pengguna/{id}', [PenggunaController::class, 'update'])->middleware(['auth','admin']);
+Route::delete('pengguna/{id}', [PenggunaController::class, 'destroy'])->middleware(['auth','admin']);
 
 //Slider
 Route::get('slider', [SliderController::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('slider/create', [SliderController::class, 'create'])->middleware(['admin']);
-Route::post('slider/store', [SliderController::class, 'store'])->name('slider.store')->middleware(['admin']);
-Route::get('slider/{id}/edit', [SliderController::class, 'edit'])->middleware(['admin']);
-Route::put('slider/{id}', [SliderController::class, 'update'])->middleware(['admin']);
-Route::delete('slider/{id}', [SliderController::class, 'destroy'])->middleware(['admin']);
+Route::get('slider/create', [SliderController::class, 'create'])->middleware(['auth','admin']);
+Route::post('slider/store', [SliderController::class, 'store'])->name('slider.store')->middleware(['auth','admin']);
+Route::get('slider/{id}/edit', [SliderController::class, 'edit'])->middleware(['auth','admin']);
+Route::put('slider/{id}', [SliderController::class, 'update'])->middleware(['auth','admin']);
+Route::delete('slider/{id}', [SliderController::class, 'destroy'])->middleware(['auth','admin']);
 
 //Grup
 Route::get('grup', [GrupController::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('grup/create', [GrupController::class, 'create'])->middleware(['admin']);
-Route::post('grup/store', [GrupController::class, 'store'])->name('grup.store')->middleware(['admin']);
-Route::get('grup/{id}/edit', [GrupController::class, 'edit'])->middleware(['admin']);
-Route::put('grup/{id}', [GrupController::class, 'update'])->middleware(['admin']);
-Route::delete('grup/{id}', [GrupController::class, 'destroy'])->middleware(['admin']);
+Route::get('grup/create', [GrupController::class, 'create'])->middleware(['auth','admin']);
+Route::post('grup/store', [GrupController::class, 'store'])->name('grup.store')->middleware(['auth','admin']);
+Route::get('grup/{id}/edit', [GrupController::class, 'edit'])->middleware(['auth','admin']);
+Route::put('grup/{id}', [GrupController::class, 'update'])->middleware(['auth','admin']);
+Route::delete('grup/{id}', [GrupController::class, 'destroy'])->middleware(['auth','admin']);
 
 //Testimoni
 use App\Http\Controllers\Toko\TestimoniController;
 Route::get('testimoni', [TestimoniController::class, 'index'])->middleware(['auth','admin-staff']);
-Route::get('testimoni/create', [TestimoniController::class, 'create'])->middleware(['admin']);
-Route::post('testimoni/store', [TestimoniController::class, 'store'])->name('testimoni.store')->middleware(['admin']);
-Route::get('testimoni/{id}/edit', [TestimoniController::class, 'edit'])->middleware(['admin']);
-Route::put('testimoni/{id}', [TestimoniController::class, 'update'])->middleware(['admin']);
-Route::delete('testimoni/{id}', [TestimoniController::class, 'destroy'])->middleware(['admin']);
+Route::get('testimoni/create', [TestimoniController::class, 'create'])->middleware(['auth','admin']);
+Route::post('testimoni/store', [TestimoniController::class, 'store'])->name('testimoni.store')->middleware(['auth','admin']);
+Route::get('testimoni/{id}/edit', [TestimoniController::class, 'edit'])->middleware(['auth','admin']);
+Route::put('testimoni/{id}', [TestimoniController::class, 'update'])->middleware(['auth','admin']);
+Route::delete('testimoni/{id}', [TestimoniController::class, 'destroy'])->middleware(['auth','admin']);
 
 //Dashboard
 Route::get('dashboard', [DashboardController::class, 'index'])->middleware(['auth','admin-staff']);
@@ -79,9 +80,8 @@ Route::get('/test', [ProdukController::class, 'index']);
 
 // Landing Page
 use App\Http\Controllers\landing\LandingController;
-
-Route::get('landingpage', [LandingController::class, 'index'])->name('landingpage');
-Route::get('landingpage/product', [LandingController::class, 'product'])->name('product');
+Route::get('/', [LandingController::class, 'index'])->name('landingpage');
+Route::get('landingpage/product', [LandingController::class, 'product'])->name('product')->middleware(['auth']);
 
 // Login
 use App\Http\Controllers\landing\LoginController;
